==> Step1_ProcessingDataset.ipynb <==
from google.colab import drive #importing drive for access
drive.mount('/content/drive') #mounting drive so colab can access files

import pandas as pd #importing pandas

# reading the csv
data = pd.read_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/chembl_TGF-betareceptortypeI.csv")
data.shape



# filter only the IC50 standard type
data_filter = data[data['standard_type'] == "IC50"]
data_filter.reset_index(drop = True, inplace = True)
data_filter.head()

# count null values
null_values = data_filter["standard_value"].isnull().sum() # sums null values
print("Nulls in standard value column: {}".format(null_values))

# removes any null values (if any)
if null_values > 0:
  print("Removing nans in standard_value column")
  print("Shape before dropping the nulls: {}".format(data_filter.shape)) # prints rows and columns
  data_filter_nans = data_filter[data_filter.standard_value.notna()]
  data_filter_nans.reset_index(drop = True, inplace = True)
  print("Shape after dropping the nulls: {}".format(data_filter_nans.shape))
else:
  print("No Nans in standard_value column")
  data_filter_nans = data_filter.copy()

# check duplicates
duplicates = data_filter_nans["canonical_smiles"].duplicated().sum()
print("Duplicates in canonical smiles column: {}".format(duplicates))

# removes any duplicates (if any)
if duplicates > 0:
  print("Removing duplicates in canonical smiles column")
  print("Shape before dropping the duplicates: {}".format(data_filter_nans.shape)) # prints rows and columns
  data_filter_dup = data_filter_nans.drop_duplicates(["canonical_smiles"])
  data_filter_dup.reset_index(drop = True, inplace = True)
  print("Shape after dropping the duplicates: {}".format(data_filter_dup.shape))
else:
  print("No duplicates in canonical smiles column")
  data_filter_dup = data_filter_nans.copy()

# count null values (canonical smiles)
null_smile = data_filter_dup["canonical_smiles"].isnull().sum() # sums null values
print("Nulls in canonical_smiles column: {}".format(null_smile))

# removes any null values (if any)
if null_smile > 0:
  print("Removing nans in canonical smiles column")
  print("Shape before dropping the nulls: {}".format(data_filter_dup.shape)) # prints rows and columns
  data_filter_smile = data_filter_dup[data_filter_dup.canonical_smiles.notna()]
  data_filter_smile.reset_index(drop = True, inplace = True)
  print("Shape after dropping the nulls: {}".format(data_filter_smile.shape))
else:
  print("No Nans in standard_value column")
  data_filter_smile = data_filter_dup.copy()

# filter out needed values
# only need canonical_smiles and standard_value (IC50 value)
select_columns = ['canonical_smiles', 'standard_value']
df = data_filter_smile[select_columns]
df.head()

# convert from IC50 to PIC50
import numpy as np

# define function for converting
def pIC50(standard_value):
  molar = standard_value * (10**-9) # converts nM to M
  return -np.log10(molar)

df['pIC50'] = df['standard_value'].apply(pIC50)
df.drop("standard_value", axis = 1, inplace = True)
df.head()


# count null values (pIC50)
new_nulls = df["pIC50"].isnull().sum() # sums null values
print("Nulls in pIC50 column: {}".format(new_nulls))

# removes any null values (if any)
if new_nulls > 0:
  print("Removing nans in pIC50 column")
  print("Shape before dropping the nulls: {}".format(df.shape)) # prints rows and columns
  df_no_null = df[df.pIC50.notna()]
  df_no_null.reset_index(drop = True, inplace = True)
  print("Shape after dropping the nulls: {}".format(df_no_null.shape))
else:
  print("No Nans in pIC50 column")
  df_no_null = df.copy()

# save dataframe
df_no_null.to_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/smile_pIC50_processed.csv", index = False)
