==> Step2_ExtractingDescriptors.ipynb <==
from google.colab import drive #importing drive for access
drive.mount('/content/drive') #mounting drive so colab can access files

! pip install rdkit

import pandas as pd #importing pandas

# reading the csv
data = pd.read_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/smile_pIC50_processed.csv")
data.head()

# extract main columns for preprocessing
MAIN_COLS = list(data.columns)
print("Main Columns", MAIN_COLS)

def extract_lipinski_and_other_descriptors(raw):
  """
  This function extracts lipinski descriptors from the raw smiles formula

  @return
  dataframe raw with each descriptor appended
  """
  from rdkit import Chem
  from rdkit.Chem import Descriptors, Lipinski
  import numpy as np
  #get formula for canonical smiles
  smile = raw['canonical_smiles']
  processed_smiles = Chem.MolFromSmiles(smile)

  #loop through descriptor functions of rdkit descriptors
  for descriptor_name, descriptor_function in Descriptors._descList:
    try:
      descriptor_value = descriptor_function(processed_smiles)
      raw[descriptor_name] = descriptor_value
    except Exception as error:
      print(str(error))
      raw[descriptor_name] = np.nan

  return raw

# run descriptor function for data
data = data.apply(extract_lipinski_and_other_descriptors, axis = 1)



print(data.shape)
data

# identify cols with missing values
missing_values = data.isnull().sum()
cols_missing = list(missing_values[missing_values > 0].index)
cols_missing

print(data.shape)

from sklearn.model_selection import train_test_split

# split the dataset
X = data.loc[:, [cols for cols in list(data.columns) if cols != "pIC50"]]
Y = data.loc[:, "pIC50"]

x_train,x_test, y_train, y_test = train_test_split(X, Y, test_size = 0.2)

print(x_train.shape)
print(y_train.shape)

# fill missing values
from sklearn.impute import SimpleImputer
import pickle

if (len(cols_missing) > 0):
  print("Filling Missing Values")
  imputer = SimpleImputer(strategy = "median").fit(x_train[cols_missing])
  # transform on train_data
  x_train[cols_missing] = imputer.transform(x_train[cols_missing])
  # transform on test_data
  x_test[cols_missing] = imputer.transform(x_test[cols_missing])

  # save impute model for later use
  with open("imputation_model", "wb") as impute_model:
    pickle.dump(imputer, impute_model)
    print("Model Saved")

from sklearn.preprocessing import StandardScaler

# if you want to scale data, set to True
SCALING = True

# get only required columns without pIC50 and smiles
REQUIRED_COLUMNS = [cols for cols in list(x_train.columns) if cols not in MAIN_COLS]
print(REQUIRED_COLUMNS)

if SCALING:
  print("Preprocessing")
  stc = StandardScaler().fit(x_train[REQUIRED_COLUMNS])
  # transform on train_data
  x_train[REQUIRED_COLUMNS] = stc.transform(x_train[REQUIRED_COLUMNS])
  # transform on test_data
  x_test[REQUIRED_COLUMNS] = stc.transform(x_test[REQUIRED_COLUMNS])

  # save scaling model for later use
  with open("scaling_model", "wb") as scaling_file:
    pickle.dump(stc, scaling_file)
    print("Model Saved")




# saving file to results folder of drive
with open("/content/drive/Shareddrives/1:1_Aiden_Chavda/Results/scaling_model", "wb") as scaling_file:
    pickle.dump(stc, scaling_file)
    print("Model Saved")

selected_columns = ["canonical_smiles", "pIC50"]

# save train and test csvs
# train
train_csv = pd.concat([x_train, y_train], axis = 1)
train_csv.reset_index(drop = True, inplace = True)
# select only smile and pIC50 for chemberta and pubchem
train_csv_smile = train_csv[selected_columns]
train_csv.drop(selected_columns[0], axis = 1, inplace = True)

# test
test_csv = pd.concat([x_test, y_test], axis = 1)
test_csv.reset_index(drop = True, inplace = True)
# select only smile and pIC50 for chemberta and pubchem
test_csv_smile = test_csv[selected_columns]
test_csv.drop(selected_columns[0], axis = 1, inplace = True)

print(test_csv.shape)
print(test_csv_smile.shape)
print(train_csv.shape)
print(train_csv_smile.shape)

# saving csvs

test_csv.to_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/test_regular.csv", index = False)
test_csv_smile.to_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/test_smile.csv", index = False)
train_csv.to_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/train_regular.csv", index = False)
train_csv_smile.to_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/train_smile.csv", index = False)
