==> Step3_TrainingAndSavingBM_WithKNNRandomForest.ipynb <==
from google.colab import drive #importing drive for access
drive.mount('/content/drive') #mounting drive so colab can access files

import pandas as pd #importing pandas

# reading the csv
data = pd.read_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/train_regular.csv")
data.head()

X = data.loc[:, [cols for cols in list(data.columns) if cols != "pIC50"]].values
Y = data.loc[:, "pIC50"].values

5 fold CV for all Models

from sklearn.model_selection import KFold
from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

def cross_validation(model, data = (X, Y), splits = 5):
  kf = KFold(n_splits = splits, shuffle=True, random_state=42)

  # Perform k-fold cross-validation
  rmse_scores = []
  r2_scores = []

  for train_index, valid_index in kf.split(data[0]):
    X_train, X_valid = data[0][train_index], data[0][valid_index]
    y_train, y_valid = data[1][train_index], data[1][valid_index]

    # Fit the KNN model on the training data
    model.fit(X_train, y_train)

    # Make predictions on the test data
    y_pred = model.predict(X_valid)

    # Calculate Mean Squared Error (MSE)
    rmse = np.sqrt(mean_squared_error(y_valid, y_pred))
    r2 = r2_score(y_valid, y_pred)
    rmse_scores.append(rmse)
    r2_scores.append(r2)

  # get arrays
  rmse_set = np.array(rmse_scores)
  r2_set = np.array(r2_scores)

  print("{}".format(rmse_set) + "|" + "{}".format(r2_set) + "|" + "{}".format(rmse_set.mean()) + "|" + "{}".format(r2_set.mean()))
  # print("RMSE Scores: {}".format(rmse_set))
  # print("R2 Scores: {}".format(r2_set))
  # print("RMSE Mean: {}".format(rmse_set.mean()))
  # print("R2 Mean: {}".format(r2_set.mean()))

from sklearn.neighbors import KNeighborsRegressor

# Use 5-Fold Cross Validation for hyper parameter tuning
# Try out diff values, find best value for parameter
for n in range(1, 21):
  # print(n, end = "")
  knn = KNeighborsRegressor(n_neighbors = n)
  cross_validation(knn)

from sklearn.ensemble import RandomForestRegressor

# Use 5-Fold Cross Validation for hyper parameter tuning
# Try out diff values, find best value for parameter

#random_forest = RandomForestRegressor(n_estimators = 10, max_depth = 5)
for n in range(10, 101, 10):
  for d in range(1, 8):
    print(str(n) + "|" + str(d) + "|", end = "")
    random_forest = RandomForestRegressor(n_estimators = n, max_depth = d)
    cross_validation(random_forest)

import pickle
import os

MODEL_SAVE_FOLDER = "/content/drive/Shareddrives/1:1_Aiden_Chavda/Results"
MODEL_NAME = "best_model_KNN"

# Model with the lowest RMSE or highest r2
best_model = KNeighborsRegressor(n_neighbors=5)

# Fit the model on the full training dataset
best_model.fit(X, Y)

fh = open(os.path.join(MODEL_SAVE_FOLDER, MODEL_NAME), "wb")
pickle.dump(best_model, fh)
fh.close()

import pickle
import os

MODEL_SAVE_FOLDER = "/content/drive/Shareddrives/1:1_Aiden_Chavda/Results"
MODEL_NAME = "best_model_random_forest"

# Model with the lowest RMSE or highest r2
best_model = RandomForestRegressor(n_estimators = 60, max_depth = 7)

# Fit the model on the full training dataset
best_model.fit(X, Y)

fh = open(os.path.join(MODEL_SAVE_FOLDER, MODEL_NAME), "wb")
pickle.dump(best_model, fh)
fh.close()
