==> Step5_TestingBM_KNN_RF.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np

# read the csv
test_data = pd.read_csv("/content/drive/Shareddrives/1:1_Aiden_Chavda/Dataset/test_regular.csv")

# seperate features and labels
x_test = test_data.loc[:, [cols for cols in list(test_data.columns) if cols != "pIC50"]].values
y_test = test_data.loc[:, "pIC50"].values

import pickle

# loading KNN best model
MODEL_PATH = "/content/drive/Shareddrives/1:1_Aiden_Chavda/Results/best_model_KNN"

file_name = open(MODEL_PATH, 'rb')
model = pickle.load(file_name)
print("Model is loaded")

from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

# get the test predictions
y_pred = model.predict(x_test)

# get metrices
mse = mean_squared_error(y_test, y_pred)
r2_val = r2_score(y_test, y_pred)

print("Test MSE: {}".format(mse))
print("Test RMSE: {}".format(np.sqrt(mse)))
print("Test r2 score: {}".format(r2_val))

# Start of testing for random forest

import pickle

# Loading random forest model

MODEL_PATH = "/content/drive/Shareddrives/1:1_Aiden_Chavda/Results/best_model_random_forest"

file_name = open(MODEL_PATH, 'rb')
model = pickle.load(file_name)
print("Model is loaded")

from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

# get the test predictions
y_pred = model.predict(x_test)

# get metrices
mse = mean_squared_error(y_test, y_pred)
r2_val = r2_score(y_test, y_pred)

print("Test MSE: {}".format(mse))
print("Test RMSE: {}".format(np.sqrt(mse)))
print("Test r2 score: {}".format(r2_val))
